+++ Plugins/ChronicleEngine/Source/ChronicleEngine/Private/ChronicleEngine.cpp
@@ -7,6 +7,7 @@ UE_DEFINE_GAMEPLAY_TAG_STATIC(TAG_Chronicle_Event_Test, "Chronicle.Event.Test");
 UE_DEFINE_GAMEPLAY_TAG_STATIC(TAG_Chronicle_Event_Async, "Chronicle.Event.Async");
 UE_DEFINE_GAMEPLAY_TAG_STATIC(TAG_Chronicle_Camera_Cut, "Chronicle.Camera.Cut");
 UE_DEFINE_GAMEPLAY_TAG_STATIC(TAG_Chronicle_Camera_Blend, "Chronicle.Camera.Blend");
+UE_DEFINE_GAMEPLAY_TAG_STATIC(TAG_Chronicle_Animation_Play, "Chronicle.Animation.Play");
 UE_DEFINE_GAMEPLAY_TAG_STATIC(TAG_Chronicle_Audio_PlayVoice, "Chronicle.Audio.PlayVoice");
 UE_DEFINE_GAMEPLAY_TAG_STATIC(TAG_Chronicle_Audio_StopVoice, "Chronicle.Audio.StopVoice");
 UE_DEFINE_GAMEPLAY_TAG_STATIC(TAG_Chronicle_Speaker_Alice, "Chronicle.Speaker.Alice");